put contact logic into hook
